Move client implementation to trait since it's all generalized
